Get input from user and create volumes.
This code gets the below input from user
project,group,user,volume type, name and size

And checks if the project exists, if it does not exist create a new project.
In the same way,checks for group and user, if exists use the same
else create a new one.
With this project,group and user information create a volume with required
volume type and size given in GB.